+++ weaver/utils.py
@@ -20,7 +20,7 @@ from concurrent.futures import ALL_COMPLETED, CancelledError, ThreadPoolExecutor
 from copy import deepcopy
 from datetime import datetime
 from types import MappingProxyType
-from typing import TYPE_CHECKING, Any, Iterable, Protocol, overload
+from typing import TYPE_CHECKING, Any, Iterable, Protocol, cast, overload
 from urllib.parse import ParseResult, parse_qsl, unquote, urlparse, urlunsplit
 
 import boto3
@@ -152,6 +152,14 @@ if TYPE_CHECKING:
         "cache": NotRequired[bool],
         "cache_enabled": NotRequired[bool],
     }, total=False)
+    RequestOptionsConfigMatcher = TypedDict("RequestOptionsConfigMatcher", {
+        "url": Required[Union[str, List[str]]],
+        "method": NotRequired[Union[AnyRequestMethod, Literal["*"], List[Union[AnyRequestMethod, Literal["*"]]]]],
+    }, total=True)
+    RequestOptionsConfigEntry = Union[RequestOptions, RequestOptionsConfigMatcher]
+    RequestOptionsSpecification = TypedDict("RequestOptionsSpecification", {
+        "requests": List[RequestOptionsConfigEntry],
+    }, total=True)
     RequestCachingKeywords = Dict[str, AnyValueType]
     RequestCachingFunction = Callable[[AnyRequestMethod, str, RequestCachingKeywords], Response]
 
@@ -1961,21 +1969,21 @@ def get_request_options(method, url, settings):
         )
         return {}
     settings = get_settings(settings)  # ensure settings, could be any container
-    req_opts_specs = settings.get("weaver.request_options", None)
+    req_opts_specs = cast("RequestOptionsSpecification", settings.get("weaver.request_options", None))
     if not isinstance(req_opts_specs, dict):
         # empty request options is valid (no file specified),
         # but none pre-processed by app means the settings come from unexpected source
         LOGGER.warning("Settings container provided by [%s] missing request options specification. "
                        "Request might not be executed with expected configuration.", get_caller_name(skip=2))
         return {}
-    request_options = {}
+    request_options = cast("RequestOptions", {})
     request_entries = req_opts_specs.get("requests", []) or []
     for req_opts in request_entries:
         req_meth = req_opts.get("method", "")
         if req_meth:
-            methods = req_meth if isinstance(req_meth, list) else [req_meth]
+            methods = req_meth if isinstance(req_meth, list) else req_meth.split(",")
             methods = [meth.upper() for meth in methods]
-            if method.upper() not in methods:
+            if method.upper() not in methods and "*" not in methods:
                 continue
         req_urls = req_opts.get("url")
         req_urls = [req_urls] if not isinstance(req_urls, list) else req_urls